+++ Application/react-native.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    assets: [
+      './assets/fonts'
+    ],
+}
